@@ -11,19 +11,19 @@
   "private": true,
   "dependencies": {
     "@angular-material-components/datetime-picker": "^7.0.1",
-    "@angular/animations": "~13.3.6",
-    "@angular/cdk": "^13.3.6",
-    "@angular/common": "~13.3.6",
-    "@angular/compiler": "~13.3.6",
-    "@angular/core": "~13.3.6",
+    "@angular/animations": "^14.0.1",
+    "@angular/cdk": "^14.0.1",
+    "@angular/common": "^14.0.1",
+    "@angular/compiler": "^14.0.1",
+    "@angular/core": "^14.0.1",
     "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "~13.3.6",
-    "@angular/localize": "~13.3.6",
-    "@angular/material": "^13.3.6",
-    "@angular/material-moment-adapter": "^13.3.6",
-    "@angular/platform-browser": "~13.3.6",
-    "@angular/platform-browser-dynamic": "~13.3.6",
-    "@angular/router": "~13.3.6",
+    "@angular/forms": "^14.0.1",
+    "@angular/localize": "^14.0.1",
+    "@angular/material": "^14.0.1",
+    "@angular/material-moment-adapter": "^14.0.1",
+    "@angular/platform-browser": "^14.0.1",
+    "@angular/platform-browser-dynamic": "^14.0.1",
+    "@angular/router": "^14.0.1",
     "@auth0/auth0-angular": "^1.9.0",
     "@fortawesome/fontawesome-free": "^6.1.1",
     "@ng-bootstrap/ng-bootstrap": "^12.0.2",
@@ -45,9 +45,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.5",
-    "@angular/cli": "~13.3.5",
-    "@angular/compiler-cli": "~13.3.6",
+    "@angular-devkit/build-angular": "^14.0.1",
+    "@angular/cli": "^14.0.1",
+    "@angular/compiler-cli": "^14.0.1",
     "@types/file-saver": "^2.0.5",
     "@types/jasmine": "~3.10.0",
     "@types/moment": "^2.13.0",
@@ -60,6 +60,6 @@
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.5.2"
+    "typescript": "~4.7.3"
   }
 }
